Added more functionality: The simulation breaks the current loop if the last computed landing spot deviates too much from the desired threshold.
Also rearranged the way it plots data in matrix to make it more readable.

+Added some documentation to make the code more comprehensive

Caution: The y-coordinates of the output seems right, such that the center row are all zero because at the point the y-component of wind is zero. I am not sure about x-coordinates since it seems like the numbers are fluctuating. I know the rocket is launched at an angle, perhaps that's the reason?